pyproject.toml
[tool.poetry]
name = "docops-agent"
version = "0.1.0"
description = "Document Ops Agent MVP (no LLM)."
authors = ["huzirui <173997453+huzirui-123@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.2.0"
pydantic = "^2.8.2"
PyYAML = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.8"
mypy = "^1.10.0"


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
files = ["core", "apps", "tests"]
disallow_untyped_defs = false
check_untyped_defs = false
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
show_error_codes = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
